137
몬스터 run 플레이어 따라오게함
몬스터 사망시 ray 제거 완료

플레이어 몬스터 크기 수정
grunt 텍스처 수정

디폴트 속도 define

+++ Engine_SOURCE/Renderer.cpp
@@ -712,8 +712,8 @@ namespace dru::renderer
 		noise->BindShaderResource(eShaderStage::CS, 16);
 
 		NoiseCB info = {};
-		info.noiseSize.x = noise->GetWidth(); // 노이즈 텍스처 사이즈를 상수버퍼로 전달해줌
-		info.noiseSize.y = noise->GetHeight();
+		info.noiseSize.x = static_cast<float>(noise->GetWidth()); // 노이즈 텍스처 사이즈를 상수버퍼로 전달해줌
+		info.noiseSize.y = static_cast<float>(noise->GetHeight());
 
 		CConstantBuffer* cb = renderer::constantBuffers[(UINT)eCBType::Noise];
 		cb->SetData(&info);
